General: Clear IDC error if mismatching URL or no longer opted in

+++ class.jetpack.php
@@ -5297,12 +5297,14 @@ p {
 		}
 
 		// Last, let's check if sync is erroring due to an IDC. If so, set the site to staging mode.
-		if ( ! $is_staging ) {
-			$sync_error = Jetpack_Options::get_option( 'sync_error_idc' );
-			if ( $sync_error && $sync_error === get_home_url() ) {
-				$is_staging = true;
-			} else if ( $sync_error ) {
+		if ( ! $is_staging && ( $sync_error = Jetpack_Options::get_option( 'sync_error_idc' ) ) ) {
+			 if ( ! self::sync_idc_optin() || $sync_error !== get_home_url() ) {
+				// If the values do not match, or the site is no longer opted in,
+				// delete the `jetpack_sync_error_idc` option
 				Jetpack_Options::delete_option( 'sync_error_idc' );
+			} else {
+				// If the value stored in $sync_error is the same as get_home_url(), then put the site in staging.
+				$is_staging = true;
 			}
 		}
